Apple의 타입을 double로 변경

+++ src/practice/Apple.java
@@ -2,9 +2,9 @@ package practice;
 
 public class Apple {
 
-		int a;
-		int b;
-		int c;
+		double a;
+		double b;
+		double c;
 		float d;
 		
 		public Apple(int a, int b, int c ) {
